+++ libyul/backends/wasm/WasmCodeTransform.h
@@ -95,6 +95,7 @@ private:
 	std::vector<wasm::VariableDeclaration> m_localVariables;
 	std::vector<wasm::GlobalVariableDeclaration> m_globalVariables;
 	std::map<YulString, wasm::FunctionImport> m_functionsToImport;
+	std::string m_functionBodyLabel;
 	std::stack<std::pair<std::string, std::string>> m_breakContinueLabelNames;
 };