+++ test/libsolidity/semanticTests/intheritance/inherited_function_through_dispatch.sol
@@ -1,0 +1,21 @@
+contract A {
+    function f() internal virtual returns (uint256) {
+        return 1;
+    }
+}
+
+
+contract B is A {
+    function f() internal override returns (uint256) {
+        return 2;
+    }
+
+    function g() public returns (uint256) {
+        function() internal returns (uint256) ptr = A.f;
+        return ptr();
+    }
+}
+// ====
+// compileViaYul: also
+// ----
+// g() -> 1