+++ test/libsolidity/syntaxTests/nameAndTypeResolution/204_overwrite_memory_location_external.sol
@@ -2,4 +2,3 @@ contract C {
     function f(uint[] memory a) external {}
 }
 // ----
-// TypeError: (28-43): Data location must be "calldata" for parameter in external function, but "memory" was given.