+++ Changelog.md
@@ -20,6 +20,8 @@ Bugfixes:
 ### 0.6.9 (unreleased)
 
 Language Features:
+ * Permit calldata location for all variables.
+ * Yul: EVM instruction `pc()` is marked deprecated and will be removed in the next breaking release.
 
 
 Compiler Features:
@@ -31,6 +33,9 @@ Compiler Features:
 Bugfixes:
  * Optimizer: Fixed a bug in BlockDeDuplicator.
  * Type Checker: Disallow assignments to storage variables of type ``mapping``.
+ * Type Checker: Disallow inline arrays of non-nameable types.
+ * Type Checker: Fix internal compiler error when accessing members of array slices.
+ * Type Checker: Fix internal compiler error when trying to decode too large static arrays.
  * NatSpec: DocString block is terminated when encountering an empty line.
  * Scanner: Fix bug when two empty NatSpec comments lead to scanning past EOL.
  * Code Generator: Trigger proper unimplemented errors on certain array copy operations.